Can run most of test suite (with hacks).
