+++ meson/mesonmain.py
@@ -220,6 +220,3 @@ def run(args):
             traceback.print_exc()
         return 1
     return 0
-
-if __name__ == '__main__':
-    sys.exit(run(sys.argv[:]))
