+++ scripts/delwithsuffix.py
