+++ scripts/gtkdochelper.py
