+++ scripts/mesonintrospect.py
