+++ scripts/regen_checker.py
