+++ scripts/symbolextractor.py
