+++ scripts/vcstagger.py
